api/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-docx==1.1.0
PyMuPDF==1.23.26
supabase==2.4.0
python-dotenv==1.0.1
requests==2.31.0